move validation funcs into util/validation
