+++ pkg/util/validation/validation.go
@@ -4,6 +4,7 @@ import (
 	"fmt"
 
 	corev1 "k8s.io/api/core/v1"
+	"k8s.io/apimachinery/pkg/util/sets"
 
 	policyv1alpha1 "github.com/karmada-io/karmada/pkg/apis/policy/v1alpha1"
 	"github.com/karmada-io/karmada/pkg/util"
@@ -29,6 +30,38 @@ func ValidatePolicyFieldSelector(fieldSelector *policyv1alpha1.FieldSelector) er
 		case corev1.NodeSelectorOpIn, corev1.NodeSelectorOpNotIn:
 		default:
 			return fmt.Errorf("unsupported operator %q, must be In or NotIn", matchExpression.Operator)
+		}
+	}
+
+	return nil
+}
+
+// ValidateSpreadConstraint tests if the constraints is valid.
+func ValidateSpreadConstraint(spreadConstraints []policyv1alpha1.SpreadConstraint) error {
+	spreadByFields := sets.NewString()
+
+	for _, constraint := range spreadConstraints {
+		// SpreadByField and SpreadByLabel should not co-exist
+		if len(constraint.SpreadByField) > 0 && len(constraint.SpreadByLabel) > 0 {
+			return fmt.Errorf("invalid constraints: SpreadByLabel(%s) should not co-exist with spreadByField(%s)", constraint.SpreadByLabel, constraint.SpreadByField)
+		}
+
+		// If MaxGroups provided, it should greater or equal than MinGroups.
+		if constraint.MaxGroups > 0 && constraint.MaxGroups < constraint.MinGroups {
+			return fmt.Errorf("maxGroups(%d) lower than minGroups(%d) is not allowed", constraint.MaxGroups, constraint.MinGroups)
+		}
+
+		if len(constraint.SpreadByField) > 0 {
+			spreadByFields.Insert(string(constraint.SpreadByField))
+		}
+	}
+
+	if spreadByFields.Len() > 0 {
+		// If one of spread constraints are using 'SpreadByField', the 'SpreadByFieldCluster' must be included.
+		// For example, when using 'SpreadByFieldRegion' to specify region groups, at the meantime, you must use
+		// 'SpreadByFieldCluster' to specify how many clusters should be selected.
+		if !spreadByFields.Has(string(policyv1alpha1.SpreadByFieldCluster)) {
+			return fmt.Errorf("the cluster spread constraint must be enabled in one of the constraints in case of SpreadByField is enabled")
 		}
 	}
